+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.2.0
+
+* Add optional item requirements
+
 # 2.1.0
 
 * Add sounds for bottling & using experience